Control input height & center search icon
In some applications the input height can be overriden if a global style is set to `input[type="text"]`.  By adding `height: inherit` to the input's styles, we ensure the height doesn't unintentionly increase.

This commit also adds `display: flex` and `align-items: center` to the search and spinner icon's parent div. This ensures those will always be vertically centered, even in cases that the input height is increased. (Intentionally or otherwise)