Refactor image container to use CSS Grids
For some reason the browsers scroll bar has started impeeding on the container that holds the grid of images. This was causing a 3 column grid to go to a 2 column grid, with a large empty space on the right.

By using CSS Grids, this simplifies the code needed to calculate how big the images should be and columns needed.

Instead of the `Photo` handling column width and gap per image, the parent container handles everything.

It sets the column sizes and dynamicly updates their width via a css variable that recieves the computed `searchResultWidth`.

This css variable sits in a `calc` to remove any column any space taken up by the browser's scroll bar.

End result is to ensure that the columns are preserved at all screen sizes and with browser scroll bars.