Move image display/size properties to a css class
I tried to control the height with CSS Grid, but the image would keep breaking out of it's container, unless a height was set on the image explicitly.

Currently the image height is calculated based on the `photoRatio` and `searchResultWidth` values. This value is then passed down via props through the `<Photo>` component to the `<SpinnerImg>` component.

In the prevoius commit we saved this calculated height into a css variable. This lets us use it in a new named class called `.unsplash-react__image`.

To help with clarity within the file, I also move the display/size properties to a named class since they are all related.

I didn't move the border propeties to this class, since those are relying on JS values. It felt beyond the scope of this PR to do further restructuring.