Style fixes for image grid and input field (#25)
* Control input height & center search icon

In some applications the input height can be overriden if a global style is set to `input[type="text"]`.  By adding `height: inherit` to the input's styles, we ensure the height doesn't unintentionly increase.

This commit also adds `display: flex` and `align-items: center` to the search and spinner icon's parent div. This ensures those will always be vertically centered, even in cases that the input height is increased. (Intentionally or otherwise)

* Refactor image container to use CSS Grids

For some reason the browsers scroll bar has started impeeding on the container that holds the grid of images. This was causing a 3 column grid to go to a 2 column grid, with a large empty space on the right.

By using CSS Grids, this simplifies the code needed to calculate how big the images should be and columns needed.

Instead of the `Photo` handling column width and gap per image, the parent container handles everything.

It sets the column sizes and dynamicly updates their width via a css variable that recieves the computed `searchResultWidth`.

This css variable sits in a `calc` to remove any column any space taken up by the browser's scroll bar.

End result is to ensure that the columns are preserved at all screen sizes and with browser scroll bars.

* Bump version: 0.2.0 -> 0.2.1

* Remove unsused `width` prop

This was removed in a prevoius commit, in favor of passing the same value to css grid to create columns.

* Capture `searchResultHeight` in a css variable

This will allow us to use the calculated height for images in css classes, instead of prop chaining it to the image.

* Move image display/size properties to a css class

I tried to control the height with CSS Grid, but the image would keep breaking out of it's container, unless a height was set on the image explicitly.

Currently the image height is calculated based on the `photoRatio` and `searchResultWidth` values. This value is then passed down via props through the `<Photo>` component to the `<SpinnerImg>` component.

In the prevoius commit we saved this calculated height into a css variable. This lets us use it in a new named class called `.unsplash-react__image`.

To help with clarity within the file, I also move the display/size properties to a named class since they are all related.

I didn't move the border propeties to this class, since those are relying on JS values. It felt beyond the scope of this PR to do further restructuring.

* Remove no longer used `height` prop

* Remove unused columns & index props